feat: aws sqs client refactor
