+++ src/requestor/clients/sqs-client.ts
@@ -6,18 +6,12 @@ import { SQSOptions } from '../../interfaces/sqs-options.interface';
 import { v4 as uuidv4 } from 'uuid';
 import { Producer } from 'sqs-producer';
 
-import * as crypto from 'crypto';
-
-import {
-  CreateQueueCommand,
-  GetQueueUrlCommand,
-  SQSClient,
-} from '@aws-sdk/client-sqs';
+import { AWSSQSClient } from '@utils/aws/sqs.client';
 
 export class ClientSQS extends ClientProxy {
   protected readonly logger: Logger;
   protected producers: Map<string, Producer>;
-  protected sqsClient: SQSClient;
+  protected awsSqsClient: AWSSQSClient;
 
   constructor(protected readonly options: SQSOptions) {
     super();
@@ -27,8 +21,6 @@ export class ClientSQS extends ClientProxy {
 
     this.producers = new Map<string, Producer>();
     this.logger = new Logger(ClientProxy.name);
-
-    this.sqsClient = new SQSClient();
   }
 
   protected publish(
@@ -45,10 +37,10 @@ export class ClientSQS extends ClientProxy {
       let queueUrl: string;
 
       try {
-        queueUrl = await this.getQueueUrl(pattern);
+        queueUrl = await this.awsSqsClient.getQueueUrl(pattern);
       } catch (error) {
         this.logger.debug(`Queue ${pattern} not found, creating it...`);
-        queueUrl = await this.createQueue(pattern);
+        queueUrl = await this.awsSqsClient.createQueue(pattern);
       }
 
       producer = Producer.create({
@@ -64,45 +56,12 @@ export class ClientSQS extends ClientProxy {
     });
   }
 
-  async getQueueUrl(pattern: string) {
-    const md5Pattern = crypto.createHash('md5').update(pattern).digest('hex');
-
-    const getQueueUrlCommand = new GetQueueUrlCommand({
-      QueueName: md5Pattern,
-    });
-
-    const { QueueUrl } = await this.sqsClient.send(getQueueUrlCommand);
-
-    this.logger.debug(`Queue URL for ${pattern} is ${QueueUrl}`);
-
-    return QueueUrl;
-  }
-
-  async createQueue(pattern: string) {
-    const md5Pattern = crypto.createHash('md5').update(pattern).digest('hex');
-
-    const createQueueCommand = new CreateQueueCommand({
-      QueueName: md5Pattern,
-      Attributes: {
-        VisibilityTimeout: '60',
-        MessageRetentionPeriod: '1209600',
-      },
-    });
-
-    const response = await this.sqsClient.send(createQueueCommand);
-
-    this.logger.debug(
-      `Creating queue ${pattern} with URL ${response.QueueUrl}`,
-    );
-
-    return response.QueueUrl;
-  }
-
   public async connect() {
-    const { sqsUri, region } = this.options;
+    const { sqsUri } = this.options;
 
     if (!sqsUri) throw new Error('Unable to init AWS SQS, missing sqsUri');
-    if (!region) throw new Error('Unable to init AWS SQS, missing region');
+
+    this.awsSqsClient = AWSSQSClient.getInstance();
   }
 
   public close() {
