+++ src/responder/transporters/server-sqs.ts
@@ -10,18 +10,11 @@ import { Consumer } from 'sqs-consumer';
 import { Message } from 'sqs-producer';
 import { isObservable } from 'rxjs';
 import { Logger } from '@nestjs/common';
-
-import * as crypto from 'crypto';
-import {
-  CreateQueueCommand,
-  GetQueueUrlCommand,
-  SQSClient,
-} from '@aws-sdk/client-sqs';
+import { AWSSQSClient } from '@utils/aws/sqs.client';
 
 export class ServerSQS extends Server implements CustomTransportStrategy {
   private consumers: Map<string, Consumer>;
-
-  private sqsClient: SQSClient;
+  private awsSqsClient: AWSSQSClient;
 
   constructor(
     private readonly options: SQSOptions,
@@ -33,8 +26,7 @@ export class ServerSQS extends Server implements CustomTransportStrategy {
     this.initializeDeserializer(options);
 
     this.consumers = new Map<string, Consumer>();
-
-    this.sqsClient = new SQSClient();
+    this.awsSqsClient = AWSSQSClient.getInstance();
   }
 
   public async listen(callback: () => void) {
@@ -60,22 +52,19 @@ export class ServerSQS extends Server implements CustomTransportStrategy {
     handler: MessageHandler<any, any, any>,
     queueName: string,
   ) {
-    const { sqsUri, waitTimeSeconds } = this.options;
-
     if (this.consumers.has(queueName)) return;
 
     let queueUrl: string;
 
     try {
-      queueUrl = await this.getQueueUrl(queueName);
+      queueUrl = await this.awsSqsClient.getQueueUrl(queueName);
     } catch (error) {
       this.logger.debug(`Queue ${queueName} not found, creating it...`);
-      queueUrl = await this.createQueue(queueName);
+      queueUrl = await this.awsSqsClient.createQueue(queueName);
     }
 
     const app = Consumer.create({
       queueUrl: queueUrl,
-      waitTimeSeconds: waitTimeSeconds ?? 0,
 
       handleMessage: async (data: unknown) => {
         const message = data as Message;
@@ -94,40 +83,6 @@ export class ServerSQS extends Server implements CustomTransportStrategy {
 
     app.start();
     this.consumers.set(queueName, app);
-  }
-
-  async getQueueUrl(pattern: string) {
-    const md5Pattern = crypto.createHash('md5').update(pattern).digest('hex');
-
-    const getQueueUrlCommand = new GetQueueUrlCommand({
-      QueueName: md5Pattern,
-    });
-
-    const { QueueUrl } = await this.sqsClient.send(getQueueUrlCommand);
-
-    this.logger.debug(`Queue URL for ${pattern} is ${QueueUrl}`);
-
-    return QueueUrl;
-  }
-
-  async createQueue(pattern: string) {
-    const md5Pattern = crypto.createHash('md5').update(pattern).digest('hex');
-
-    const createQueueCommand = new CreateQueueCommand({
-      QueueName: md5Pattern,
-      Attributes: {
-        VisibilityTimeout: '60',
-        MessageRetentionPeriod: '1209600',
-      },
-    });
-
-    const response = await this.sqsClient.send(createQueueCommand);
-
-    this.logger.debug(
-      `Creating queue ${pattern} with URL ${response.QueueUrl}`,
-    );
-
-    return response.QueueUrl;
   }
 
   public close() {
